Add logout api

+++ src/components/layout/user/AvatarMenu.tsx
@@ -2,6 +2,7 @@ import { Menu, MenuItem } from "@mui/material";
 import React, { useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import { UserContext } from "../../../App";
+import { authServices } from "../../../utils/api/auth/services";
 import CustomizedDialogPassword from "./ResetPasswordDailog";
 import CustomizedDialogStack from "./StackDailog";
 
@@ -9,16 +10,18 @@ export function AvatarMenu({ anchorEl, open, handleClose }: any) {
   const { setIsSignedIn, setUser } = useContext(UserContext);
 
   const nav = useNavigate()
-  const logout = () => {
-    localStorage.clear();
-    nav("/")
-    if (setUser) {
-      setUser(null);
+  const logout = async() => {
+    const { response } = await authServices.logout();
+    if (response) {
+      localStorage.clear();
+      nav("/")
+      if (setUser) {
+        setUser(null);
+      }
+      if (setIsSignedIn) {
+        setIsSignedIn(false);
+      }
     }
-    if (setIsSignedIn) {
-      setIsSignedIn(false);
-    }
-    
   };
   return (
     <>
